chore: move `data-badge` stories to `data-badge-docs` package
