+++ packages/docs/data-badge-docs/package.json
@@ -21,5 +21,9 @@
   "publishConfig": {
     "access": "public",
     "provenance": true
+  },
+  "devDependencies": {
+    "@nl-design-system-candidate/data-badge-react": "workspace:*",
+    "@storybook/react-vite": "9.1.5"
   }
 }
